+++ src/PhpSpreadsheet/Calculation/LookupRef.php
@@ -464,9 +464,10 @@ class LookupRef
      *
      * @param mixed $lookupValue The value that you want to match in lookup_array
      * @param mixed $lookupArray The range of cells being searched
-     * @param mixed $matchType The number -1, 0, or 1. -1 means above, 0 means exact match, 1 means below. If match_type is 1 or -1, the list has to be ordered.
+     * @param mixed $matchType The number -1, 0, or 1. -1 means above, 0 means exact match, 1 means below.
+     *                         If match_type is 1 or -1, the list has to be ordered.
      *
-     * @return int The relative position of the found item
+     * @return int|string The relative position of the found item
      */
     public static function MATCH($lookupValue, $lookupArray, $matchType = 1)
     {
@@ -474,9 +475,10 @@ class LookupRef
         $lookupValue = Functions::flattenSingleValue($lookupValue);
         $matchType = ($matchType === null) ? 1 : (int) Functions::flattenSingleValue($matchType);
 
-        $initialLookupValue = $lookupValue;
-        // MATCH is not case sensitive
-        $lookupValue = StringHelper::strToLower($lookupValue);
+        // MATCH is not case sensitive, so we convert lookup value to be lower cased in case it's string type.
+        if (is_string($lookupValue)) {
+            $lookupValue = StringHelper::strToLower($lookupValue);
+        }
 
         // Lookup_value type has to be number, text, or logical values
         if ((!is_numeric($lookupValue)) && (!is_string($lookupValue)) && (!is_bool($lookupValue))) {
@@ -522,16 +524,54 @@ class LookupRef
         // find the match
         // **
 
-        if ($matchType == 0 || $matchType == 1) {
+        if ($matchType === 0 || $matchType === 1) {
             foreach ($lookupArray as $i => $lookupArrayValue) {
-                $onlyNumeric = is_numeric($lookupArrayValue) && is_numeric($lookupValue);
-                $onlyNumericExactMatch = $onlyNumeric && $lookupArrayValue == $lookupValue;
-                $nonOnlyNumericExactMatch = !$onlyNumeric && $lookupArrayValue === $lookupValue;
-                $exactMatch = $onlyNumericExactMatch || $nonOnlyNumericExactMatch;
-                if (($matchType == 0) && $exactMatch) {
-                    //    exact match
-                    return $i + 1;
-                } elseif (($matchType == 1) && ($lookupArrayValue <= $lookupValue)) {
+                $typeMatch = gettype($lookupValue) === gettype($lookupArrayValue);
+                $exactTypeMatch = $typeMatch && $lookupArrayValue === $lookupValue;
+                $nonOnlyNumericExactMatch = !$typeMatch && $lookupArrayValue === $lookupValue;
+                $exactMatch = $exactTypeMatch || $nonOnlyNumericExactMatch;
+
+                if ($matchType === 0) {
+                    if ($typeMatch && is_string($lookupValue) && (bool) preg_match('/([\?\*])/', $lookupValue)) {
+                        $splitString = $lookupValue;
+                        $chars = array_map(function ($i) use ($splitString) {
+                            return mb_substr($splitString, $i, 1);
+                        }, range(0, mb_strlen($splitString) - 1));
+
+                        $length = count($chars);
+                        $pattern = '/^';
+                        for ($j = 0; $j < $length; ++$j) {
+                            if ($chars[$j] === '~') {
+                                if (isset($chars[$j + 1])) {
+                                    if ($chars[$j + 1] === '*') {
+                                        $pattern .= preg_quote($chars[$j + 1], '/');
+                                        ++$j;
+                                    } elseif ($chars[$j + 1] === '?') {
+                                        $pattern .= preg_quote($chars[$j + 1], '/');
+                                        ++$j;
+                                    }
+                                } else {
+                                    $pattern .= preg_quote($chars[$j], '/');
+                                }
+                            } elseif ($chars[$j] === '*') {
+                                $pattern .= '.*';
+                            } elseif ($chars[$j] === '?') {
+                                $pattern .= '.{1}';
+                            } else {
+                                $pattern .= preg_quote($chars[$j], '/');
+                            }
+                        }
+
+                        $pattern .= '$/';
+                        if ((bool) preg_match($pattern, $lookupArrayValue)) {
+                            // exact match
+                            return $i + 1;
+                        }
+                    } elseif ($exactMatch) {
+                        // exact match
+                        return $i + 1;
+                    }
+                } elseif (($matchType === 1) && $typeMatch && ($lookupArrayValue <= $lookupValue)) {
                     $i = array_search($i, $keySet);
 
                     // The current value is the (first) match
@@ -539,26 +579,26 @@ class LookupRef
                 }
             }
         } else {
-            // matchType = -1
-
-            // "Special" case: since the array it's supposed to be ordered in descending order, the
-            // Excel algorithm gives up immediately if the first element is smaller than the searched value
-            if ($lookupArray[0] < $lookupValue) {
-                return Functions::NA();
-            }
-
             $maxValueKey = null;
 
             // The basic algorithm is:
             // Iterate and keep the highest match until the next element is smaller than the searched value.
             // Return immediately if perfect match is found
             foreach ($lookupArray as $i => $lookupArrayValue) {
-                if ($lookupArrayValue == $lookupValue) {
+                $typeMatch = gettype($lookupValue) === gettype($lookupArrayValue);
+                $exactTypeMatch = $typeMatch && $lookupArrayValue === $lookupValue;
+                $nonOnlyNumericExactMatch = !$typeMatch && $lookupArrayValue === $lookupValue;
+                $exactMatch = $exactTypeMatch || $nonOnlyNumericExactMatch;
+
+                if ($exactMatch) {
                     // Another "special" case. If a perfect match is found,
                     // the algorithm gives up immediately
                     return $i + 1;
-                } elseif ($lookupArrayValue >= $lookupValue) {
+                } elseif ($typeMatch & $lookupArrayValue >= $lookupValue) {
                     $maxValueKey = $i + 1;
+                } elseif ($typeMatch & $lookupArrayValue < $lookupValue) {
+                    //Excel algorithm gives up immediately if the first element is smaller than the searched value
+                    break;
                 }
             }